Destroying Cards Requirement
Destroying Cards can only be a Requirement.
Destroying a card from the deck is the first
viable card but from other places you get to
choose. Creating a destruction effect with
potentially more (negative) scale than the extra
effect punishment is a bad idea. Therefore
destroying is only a Requirement. Might make a
banish top card of Deck/Graveyard and discard top
card of deck.

Wrote the "Show" for Card (and so monster and
spell)