Added the option to check the opponent's field.
Added a new round action to see more about the opponent's field. In a
game where you have no clue what the enemy deck is, being able to see
the cards they summon is super valuable.

Changed RoundAction to be enum so that allRoundActions does not need to
be hardcoded.