Three changes around attacking.
Only monsters on the field can attack. Previously
you could use high power monster cards with
impossible summon requirements but could still
attack on draw or discard.
The "On Victory" spells on a tapped monster can
still be triggered. This allows you to have an
on tap: attack spell and a victory spell on the
same monster.

It now costs 30 instead of 20 to attack directly.