Scale is now a monad?!
Scale is a reader with the deck and whether or not it's inside a monster
card. This allows effects/requirements etc to take the entire deck into
account but this isn't used by anything YET.
It is an except monad too with a few kinds of different ways a card can
be illegal. This means that partway through testing the legality of a
deck it could fail early and display a reason.

There are bugs where scale isn't calculated correctly at the moment so
that needs to be fixed next.