REDUCE INFINITY OMG IS IT REAL?!
Reasoning: inifinity is monster only, so the scale
limit of inifinity spells is 15. Meaning the rest
of the card must be -5, which is as far as I can
tell enough to avoid all "better for you than the
opponent" spells.

Examples:
`inifinity discard field: draw` discard triggers
are expensive meaning you're unlikely to be
getting a powerful on discard effect on the field
for free.
`infinity take 1, discard hand: draw` this one is
actually pretty powerful, unlimited hand filtering
but you take damage each time is... good... but it
is scale 15 so.....
`infinity discard, pop: drill f"useful ondiscard"`
lets you play all the useful `on discard` effects
from your deck, although it's a lot of damage.
This could be brutal in a "turing" format.

I wanna experiment with this! I can undo it later
but I think we need more infinity effects.